requirements.txt
psutil==5.9.5
py-cpuinfo==9.0.0
GPUtil==1.4.0
supabase==1.0.3
